Mask full LSN and checksum in tests
The old LSN masking didn't consider LSNs with a log id with two or more
digits. Checksums were hard-coded to 0000 so far, but the server might
actually write them.

To not clutter the test files with ever-longer sed commands, move the
command to a new sed.sh file.

Test files updated to follow this change, but there are pre-existing
problems on older PG majors and 32-bit that are still open.

Close #37.

Reported-By: Maksim Melnikov <m.melnikov@postgrespro.ru>